+++ postgresql_audit/flask.py
@@ -3,8 +3,7 @@ from __future__ import absolute_import
 from contextlib import contextmanager
 from copy import copy
 
-from flask import g, request
-from flask.globals import _app_ctx_stack, _request_ctx_stack
+from flask import g, has_request_context, request
 
 from .base import VersioningManager as BaseVersioningManager
 
@@ -14,7 +13,7 @@ class VersioningManager(BaseVersioningManager):
 
     def get_transaction_values(self):
         values = copy(self.values)
-        if context_available() and hasattr(g, 'activity_values'):
+        if has_request_context() and hasattr(g, 'activity_values'):
             values.update(g.activity_values)
         if (
             'client_addr' not in values and
@@ -33,7 +32,7 @@ class VersioningManager(BaseVersioningManager):
         from flask_login import current_user
 
         # Return None if we are outside of request context.
-        if not context_available():
+        if not has_request_context():
             return
 
         try:
@@ -44,16 +43,9 @@ class VersioningManager(BaseVersioningManager):
     @property
     def default_client_addr(self):
         # Return None if we are outside of request context.
-        if not context_available():
+        if not has_request_context():
             return
         return request.remote_addr or None
-
-
-def context_available():
-    return (
-        _app_ctx_stack.top is not None and
-        _request_ctx_stack.top is not None
-    )
 
 
 def merge_dicts(a, b):
@@ -64,7 +56,7 @@ def merge_dicts(a, b):
 
 @contextmanager
 def activity_values(**values):
-    if not context_available():
+    if not has_request_context():
         return
     if hasattr(g, 'activity_values'):
         previous_value = g.activity_values